api/requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv==1.0.0
bleach==6.2.0
argon2-cffi==23.1.0 
python-jose==3.4.0
pydantic
pydantic[email]
authlib
itsdangerous